+++ packages/data-prefetch/src/react/utils.ts
@@ -1,0 +1,11 @@
+import { useEffect, useRef } from 'react';
+
+export const useFirstMounted = (): boolean => {
+  const ref = useRef(true);
+
+  useEffect(() => {
+    ref.current = false;
+  }, []);
+
+  return ref.current;
+};