+++ packages/enhanced/jest.embed.ts
@@ -24,7 +24,7 @@ if (swcJestConfig.swcrc === undefined) {
 // swcJestConfig.module.noInterop = false;
 
 export default {
-  displayName: 'enhanced',
+  displayName: 'enhanced-experiments',
   preset: '../../jest.preset.js',
   transform: {
     '^.+\\.[tj]s$': ['@swc/jest', swcJestConfig],