+++ packages/enhanced/CHANGELOG.md
@@ -1,5 +1,33 @@
 # [0.2.0-canary.5](https://github.com/module-federation/core/compare/enhanced-0.2.0-canary.4...enhanced-0.2.0-canary.5) (2023-11-20)
 
+## 0.6.6
+
+### Patch Changes
+
+- Updated dependencies [35aead4]
+  - @module-federation/dts-plugin@0.6.6
+  - @module-federation/manifest@0.6.6
+  - @module-federation/rspack@0.6.6
+  - @module-federation/sdk@0.6.6
+  - @module-federation/runtime-tools@0.6.6
+  - @module-federation/managers@0.6.6
+  - @module-federation/bridge-react-webpack-plugin@0.6.6
+  - @module-federation/data-prefetch@0.6.6
+
+## 0.6.5
+
+### Patch Changes
+
+- Updated dependencies [3070e26]
+  - @module-federation/data-prefetch@1.0.16
+  - @module-federation/rspack@0.6.5
+  - @module-federation/sdk@0.6.5
+  - @module-federation/runtime-tools@0.6.5
+  - @module-federation/managers@0.6.5
+  - @module-federation/manifest@0.6.5
+  - @module-federation/dts-plugin@0.6.5
+  - @module-federation/bridge-react-webpack-plugin@0.6.5
+
 ## 0.6.4
 
 ### Patch Changes