+++ packages/sdk/CHANGELOG.md
@@ -1,5 +1,9 @@
 # [1.1.0-canary.1](https://github.com/module-federation/core/compare/sdk-1.0.0...sdk-1.1.0-canary.1) (2023-12-05)
 
+## 0.6.6
+
+## 0.6.5
+
 ## 0.6.4
 
 ## 0.6.3