+++ packages/third-party-dts-extractor/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @module-federation/third-party-dts-extractor
 
+## 0.6.6
+
+## 0.6.5
+
 ## 0.6.4
 
 ## 0.6.3